+++ packages/ui/src/views/user-management/components/dialogs/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useDialogHandlers'
+export * from './useDialogData'
