+++ packages/ui/src/views/user-management/components/dialogs/hooks/useDialogData/index.ts
@@ -1,0 +1,1 @@
+export * from './useDialogData'
